+++ generators/client/files.js
@@ -18,7 +18,7 @@
  */
 const mkdirp = require('mkdirp');
 const constants = require('generator-jhipster/generators/generator-constants');
-const utils = require('./utils');
+const utils = require('../utils');
 
 const MAIN_SRC_DIR = constants.CLIENT_MAIN_SRC_DIR;
 const TEST_SRC_DIR = constants.CLIENT_TEST_SRC_DIR;
